folio update (#426)
Small update ...

(1) I included the trader / user_id that Andrew added to the fill object
(2) I updated the schema of the table we can now store source,
account_id, user_id and fee as per the request from Ben
(3) I will need to find a way to update the source via the API so that
we have algo information. Ben wants this added but in the next pass
(4) Source defaults to GUI at the moment and fees are 0. account_id
doesn't seem to be used with crypto but I think maybe that will be valid
on the retail side.
(5) Handle the case that Andrew pointed out wrt realtime fills and folio
fills. We only upsert of we have a valid trader / user_id. For all other
cases we insert and do nothing on a collision.
(6) the last file updated on a cargo +nightly fmt command

---------

Co-authored-by: Andrew Lee <126016409+andrew-lee-architect@users.noreply.github.com>