Add account filter to Folio account summary requests; make gql use it in certain contexts (#465)
Right now when we query for account summaries in the retail-futures
environment, we just return all accounts (Very Bad). This upgrades Folio
types to support a filter on accounts, which gql (since it knows (ish)
which users have access to which accounts) will respect.

A follow-on PR will add more stringent checks, and improve AccountMaster
so we don't have to do some hacky stuff that we did here.