Account manager (3/6)--default account permissions (#485)
I realized the semantics of account permissions weren't complete. The
new semantics are as follows (some of which is a restatement of status
quo, for clarity):

- Accounts exist with respect to cptys--they are named (in our
terminology, mapped) by them; accounts exist whether or not cpty's have
mapped them out for Architect
- Accounts have some set of default per-user permissions which are
considered intrinsic to the account (e.g. for CQG there is a notion of
an owner)
- Additional per-user permissions may be layered on top of the default
ones; _these_ permissions are created, stored, and maintained by
Architect

As a corollary, this changes AccountPermissions to be a struct with
individual fields of type Option<bool>