Deribit reconnect logic and bug fixes (#421)
(1) There was no logic for the deribit cpty to reconnect if the socket
disconnected. This happens in the simulation environment a few times per
week. I was finally able to observe this event today and verified
everything worked as expected.

(2) Since we use the same account in canary and my local environment
when we trade in each environment it would cause the other one to crash.
I added better error handling around unsolicited cancels to stop this
from happening. I was able to test this between the two environments and
it is working as expected now.

(3) Because we are using async request handlers to send requests from
architect to the deribit websocket when the socket is not connected they
were just sitting in a queue and never processed. The UI and upstream
components saw these requests as PENDING and they never changed. So as
part of the fix for (1) we will now drain the queue and send rejects
back upstream with an appropriate message. We would never want to submit
these orders on a reconnect because they would be stale and most likely
there would be cancels requests behind them.