Jenr/cqg account proxy (#534)
allow for the CQG counterparty to proxy accounts, allowing
architect-permissioned access to CQG accounts. This should allow us to
add a `UNIQUE` constraint on the `cqg_account_id` and `cqg_trader_id`
fields in UserDB. This likely has performance bugs, and it's quite
brittle, as loading accounts and setting up the proxies only happens
once. A follow up PR will move the account loading and proxying
functionality to a loop in `run`, so that we pick up new CQG accounts as
they're created. A follow up follow up PR will make the account proxy
config a separate table in UserDB so we can pick up proxy changes on the
fly as well

---------

Co-authored-by: Andrew Lee <126016409+andrew-lee-architect@users.noreply.github.com>