Add support for algo modify messages (#613)
# Description

This PR adds a new parameter to the generic algo framework to represent
a modification message.
This means that each algo chooses what modifications they support. 

In order to make this work, you'll notice a couple other things change.
- AlgoState is now required to maintain ownership of the AlgoOrder since
it can be changed. Then the generic container and each algo itself needs
to reach into the controller state to ask for the algo order.
- rx_child_algo_msgs renamed to rx_algo_msgs because it's no longer just
child algo msgs :)

## Testing
Need to write some tests for this. Still not sure how to approach.

## Changelog

- [x] Update `CHANGELOG.md` to reflect the changes in this PR