improve type hints for architect_py grpc creation (#930)
# Description
improve type hints for architect_py grpc creation


## JsonSchema Instructions:

Please ensure that there are no duplicate types in the schema.json.

Enums *must* have a tag. Generally the tag should be `t`, but you can
communicate that in the enum description. Refer to api/DICTIONARY.md for
common titles.

The format should be as such for the enum:
#[schemars(description = "<api>tag: t, any_additional_key:
any_additional_info</api>")]

For each field, the title should be the *enum variant name|type_name*.
This makes code-gen for the Python API easier.

For example:
```rust
#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
#[serde(tag = "t")]  // the tag is here
#[schemars(description = "<api>tag: t, any_additional_key: any_additional_info</api>")]
pub enum L2BookUpdate {
    #[serde(rename = "s")]
    #[schemars(title = "Snapshot|L2BookSnapshot")]  // enum_variant_name|type_name
    Snapshot(L2BookSnapshot),
    #[serde(rename = "d")]
    #[schemars(title = "Diff|L2BookDiff")]
    Diff(L2BookDiff),
}
```